Add support for new runtests binary
A new binary which can be used to automatically download and
setup gocaves when running tests, which will run the tests against
caves.

Why a new binary and not just use standard built in testing support?
Rust integration testing runs test files as their own suites with
no way to do a setup and teardown before and after running all
tests.

Why not put all tests in a single file then?
That could get unwieldly quickly, and tests can still be run in the
standard way against clusters (real, caves, or otherwise). This
approach is intended to be used primarily for purposes such as
commit validation, providing a lightweight way to run tests without
requiring a real couchbase server cluster.