Rewrite the client crate
The client crate was getting a little inconsistent and weird to use.
Split the client into kv client and http client, both of which
also consume a http_handler.
Split the 3 clients into their own files along with the http
handler and errors.
Made http_handler operate async to match the internal kv handler.
Move runtime instantiation for both http and kv ops so that
instantiation is done at the top level of the functions rather
than at the callee or inner most levels.