Improve doc command error messages
Rework client package to not have references to nushell, making
wrapping any ClientErrors the responsibility of the cli package.
Add message() and expanded_message() to ClientError to facilitate
conversion from ClientError to ShellError.
Improve the messages and error details for ClientError.
Add a halt-on-error flag to doc commands to facilitate scripting
where error codes might be wanted.